29/8/26 Lift-based tier promotion in the pipeline, structured logging, prediction snapshots" -m "Scoring: move the lift adjustment out of report_builder and into stage4 as engineer_stock_bucket_lift + engineer_lift_adjusted_bucket. Lift is now computed causally (shift(1)-before-expanding, shrunk toward the market baseline by LIFT_PRIOR_STRENGTH) instead of from a stock's whole history, and it promotes the tier as a reclassification rather than rescaling the score — blending the two was measured and rejected (top-decile lift 3.70x -> 2.98x). Every surface now agrees with the PDFs. Bucket floors and lift gates moved into config; the structural bucket and lift are cached for incremental runs.
Logging: new utilities/logging_utilities.py with LOG_LEVEL and per-library caps;
prints replaced with loggers across ingestion, cron, and pipeline.

Ingestion: yfinance earnings-date and calendar fetches run concurrently with
jitter, DB writes stay sequential; EARNINGS_RECHECK_WINDOW_DAYS makes the recheck
window configurable, and calendar validation now fetches once per unique stock.

Outputs: stage5 writes full_df.parquet up front and passes the df to downstream
generators; new pipeline/save_predictions.py snapshots each run's upcoming calls
with MODEL_VERSION and the git commit so backtests can attribute a call.

Testing: calibration gains a baseline row, Wilson CIs, per-tier year-by-year
stability, and a score/bucket consistency check, plus explicit leakage and
in-sample caveats; new pytest suite (testing/test_pipeline.py),
weekly_prediction_quality.py, inspect_db.py. testing_results/ now gitignored.

Scripts: monday_workflow.sh replaced by scripts/full_workflow.sh.

+++ feature_engineering/post_earnings_stock_features.py
@@ -67,9 +67,9 @@ def engineer_reaction_class(df):
         Engineer is_up,is_down,is_nochange features:
         if DEFAULT_REACTION_WINDOW > REACTION_THRESHOLD
     """
-    df["is_up"] = (df[DEFAULT_REACTION_WINDOW] > REACTION_THRESHOLD ).astype(int)
-    df["is_down"] = (df[DEFAULT_REACTION_WINDOW] < - REACTION_THRESHOLD ).astype(int)
-    df["is_nochange"] = ( df[DEFAULT_REACTION_WINDOW].abs() <= REACTION_THRESHOLD ).astype(int)
+    df["is_up"] = (df[DEFAULT_REACTION_WINDOW] > REACTION_THRESHOLD ).astype("int8")
+    df["is_down"] = (df[DEFAULT_REACTION_WINDOW] < - REACTION_THRESHOLD ).astype("int8")
+    df["is_nochange"] = ( df[DEFAULT_REACTION_WINDOW].abs() <= REACTION_THRESHOLD ).astype("int8")
     return df
 
 def engineer_reaction_std(df):
